move phase start stuff to StartPhase
