+++ src/main/java/ti4/commands/game/StartPhase.java
@@ -1,20 +1,48 @@
 package ti4.commands.game;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.StringJoiner;
+
+import net.dv8tion.jda.api.entities.emoji.Emoji;
 import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import ti4.buttons.Buttons;
+import ti4.commands.cardsac.PlayAC;
+import ti4.commands.cardspn.PlayPN;
 import ti4.commands.player.TurnEnd;
+import ti4.commands.player.TurnStart;
+import ti4.commands.status.Cleanup;
 import ti4.commands.status.ListPlayerInfoButton;
+import ti4.commands.status.ListTurnOrder;
+import ti4.generator.MapGenerator;
+import ti4.generator.Mapper;
 import ti4.helpers.AgendaHelper;
+import ti4.helpers.AliasHandler;
 import ti4.helpers.ButtonHelper;
+import ti4.helpers.ButtonHelperAbilities;
+import ti4.helpers.ButtonHelperActionCards;
+import ti4.helpers.ButtonHelperFactionSpecific;
 import ti4.helpers.Constants;
+import ti4.helpers.DisplayType;
+import ti4.helpers.Emojis;
+import ti4.helpers.FoWHelper;
 import ti4.helpers.Helper;
 import ti4.map.Game;
 import ti4.map.GameSaveLoadManager;
+import ti4.map.Leader;
+import ti4.map.Planet;
 import ti4.map.Player;
+import ti4.map.Tile;
+import ti4.map.UnitHolder;
 import ti4.message.MessageHelper;
+import ti4.model.PromissoryNoteModel;
 
 public class StartPhase extends GameSubcommandData {
     public StartPhase() {
@@ -32,16 +60,12 @@ public class StartPhase extends GameSubcommandData {
 
     public static void startPhase(GenericInteractionCreateEvent event, Game game, String phase) {
         switch (phase) {
-            case "strategy" -> ButtonHelper.startStrategyPhase(event, game);
+            case "strategy" -> startStrategyPhase(event, game);
             case "voting" -> AgendaHelper.startTheVoting(game);
             case "finSpecial" -> ButtonHelper.fixAllianceMembers(game);
             case "shuffleDecks" -> game.shuffleDecks();
             case "publicObj" -> ListPlayerInfoButton.displayerScoringProgression(game, true, event.getMessageChannel(), "both");
             case "publicObjAll" -> ListPlayerInfoButton.displayerScoringProgression(game, false, event.getMessageChannel(), "1");
-            // case "unleashTheNames" -> OtherStats.sendAllNames(event);
-            // case "unleashTheNamesDS" -> OtherStats.sendAllNames(event, true, false);
-            // case "unleashTheNamesAbsol" -> OtherStats.sendAllNames(event, false, true);
-            // case "unleashTheNamesEnded" -> OtherStats.showGameLengths(event, 120);
             case "ixthian" -> AgendaHelper.rollIxthian(game, false);
             case "gameTitles" -> ButtonHelper.offerEveryoneTitlePossibilities(game);
             case "giveAgendaButtonsBack" -> Helper.giveMeBackMyAgendaButtons(game);
@@ -71,15 +95,475 @@ public class StartPhase extends GameSubcommandData {
                     MessageHelper.sendMessageToChannel(event.getMessageChannel(), endOfGameSummary);
                 }
             }
-            case "statusHomework" -> ButtonHelper.startStatusHomework(event, game);
+            case "statusHomework" -> startStatusHomework(event, game);
             case "agendaResolve" -> AgendaHelper.resolveTime(game, null);
             case "pbd1000decks" -> {
                 game.pbd1000decks();
                 GameSaveLoadManager.saveMap(game, event);
             }
-            case "action" -> ButtonHelper.startActionPhase(event, game);
+            case "action" -> startActionPhase(event, game);
             case "playerSetup" -> ButtonHelper.offerPlayerSetupButtons(event.getMessageChannel(), game);
             default -> MessageHelper.sendMessageToChannel(event.getMessageChannel(), "Could not find phase: `" + phase + "`");
+        }
+    }
+
+    public static void startStrategyPhase(GenericInteractionCreateEvent event, Game game) {
+        for (Player p2 : game.getRealPlayers()) {
+            if (game.getStoredValue("LastMinuteDeliberation") != null
+                && game.getStoredValue("LastMinuteDeliberation").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("last_minute_deliberation")) {
+                PlayAC.playAC(event, game, p2, "last_minute_deliberation", game.getMainGameChannel());
+                return;
+            }
+            if (game.getStoredValue("SpecialSession") != null
+                && game.getStoredValue("SpecialSession").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("special_session")) {
+                PlayAC.playAC(event, game, p2, "special_session", game.getMainGameChannel());
+                return;
+            }
+
+        }
+        int round = game.getRound();
+        if (game.isHasHadAStatusPhase()) {
+            round++;
+            game.setRound(round);
+        }
+        MessageHelper.sendMessageToChannel(event.getMessageChannel(), "Started Round " + round);
+        if (game.isShowBanners()) {
+            MapGenerator.drawPhaseBanner("strategy", round, game.getActionsChannel());
+        }
+        if (game.getRealPlayers().size() == 6) {
+            game.setStrategyCardsPerPlayer(1);
+        }
+        ButtonHelperFactionSpecific.checkForNaaluPN(game);
+        for (Player p2 : game.getRealPlayers()) {
+            if (game.getStoredValue("Summit") != null
+                && game.getStoredValue("Summit").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("summit")) {
+                PlayAC.playAC(event, game, p2, "summit", game.getMainGameChannel());
+            }
+
+            if (game.getStoredValue("Investments") != null
+                && game.getStoredValue("Investments").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("investments")) {
+                PlayAC.playAC(event, game, p2, "investments", game.getMainGameChannel());
+            }
+
+            if (game.getStoredValue("PreRevolution") != null
+                && game.getStoredValue("PreRevolution").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("revolution")) {
+                PlayAC.playAC(event, game, p2, "revolution", game.getMainGameChannel());
+            }
+            if (game.getStoredValue("Deflection") != null
+                && game.getStoredValue("Deflection").contains(p2.getFaction())
+                && p2.getActionCards().containsKey("deflection")) {
+                PlayAC.playAC(event, game, p2, "deflection", game.getMainGameChannel());
+            }
+            if (p2.hasLeader("zealotshero") && p2.getLeader("zealotshero").get().isActive()) {
+                if (!game.getStoredValue("zealotsHeroTechs").isEmpty()) {
+                    String list = game.getStoredValue("zealotsHeroTechs");
+                    List<Button> buttons = new ArrayList<>();
+                    for (String techID : list.split("-")) {
+                        buttons.add(Buttons.green("purgeTech_" + techID, "Purge " + Mapper.getTech(techID).getName()));
+                    }
+                    String msg = p2.getRepresentation(true, true) + " due to Saint Binal, the Rhodun hero, you have to purge 2 techs. Use buttons to purge ";
+                    MessageHelper.sendMessageToChannelWithButtons(p2.getCorrectChannel(), msg + "the first tech.", buttons);
+                    MessageHelper.sendMessageToChannelWithButtons(p2.getCorrectChannel(), msg + "the second tech.", buttons);
+                    p2.removeLeader("zealotshero");
+                    game.setStoredValue("zealotsHeroTechs", "");
+                }
+            }
+        }
+        if (!game.getStoredValue("agendaConstitution").isEmpty()) {
+            game.setStoredValue("agendaConstitution", "");
+            for (Player p2 : game.getRealPlayers()) {
+                for (String planet : p2.getPlanets()) {
+                    if (planet.contains("custodia") || planet.contains("ghoti")) {
+                        continue;
+                    }
+                    if (game.getTileFromPlanet(planet) == p2.getHomeSystemTile()) {
+                        p2.exhaustPlanet(planet);
+                    }
+                }
+            }
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(),
+                "# Exhausted all home systems due to that one agenda");
+        }
+        if (!game.getStoredValue("agendaArmsReduction").isEmpty()) {
+            game.setStoredValue("agendaArmsReduction", "");
+            for (Player p2 : game.getRealPlayers()) {
+                for (String planet : p2.getPlanets()) {
+                    if (planet.contains("custodia") || planet.contains("ghoti")) {
+                        continue;
+                    }
+                    if (ButtonHelper.isPlanetTechSkip(planet, game)) {
+                        p2.exhaustPlanet(planet);
+                    }
+                }
+            }
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "# Exhausted all tech skip planets due to that one agenda.");
+        }
+        if (!game.getStoredValue("agendaChecksNBalancesAgainst").isEmpty()) {
+            game.setStoredValue("agendaChecksNBalancesAgainst", "");
+            for (Player p2 : game.getRealPlayers()) {
+                String message = p2.getRepresentation() + " Click the names of up to 3 planets you wish to ready after Checks and Balances resolved against.";
+                List<Button> buttons = Helper.getPlanetRefreshButtons(event, p2, game);
+                buttons.add(Buttons.red("deleteButtons_spitItOut", "Done Readying Planets")); // spitItOut
+                MessageHelper.sendMessageToChannel(p2.getCardsInfoThread(), message, buttons);
+            }
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(),
+                "# Sent buttons to refresh 3 planets due to Checks and Balances.");
+        }
+        if (!game.getStoredValue("agendaRevolution").isEmpty()) {
+            game.setStoredValue("agendaRevolution", "");
+            for (Player p2 : game.getRealPlayers()) {
+                String message = p2.getRepresentation() + " Exhaust 1 planet for each tech you own (" + p2.getTechs().size() + ")";
+
+                List<Button> buttons = Helper.getPlanetExhaustButtons(p2, game);
+                buttons.add(Buttons.red("deleteButtons_spitItOut", "Done Exhausting")); // spitItOut
+                MessageHelper.sendMessageToChannel(p2.getCardsInfoThread(), message, buttons);
+            }
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "# Sent buttons to exhaust 1 planet for each tech due to Anti-Intellectual Revolution resolving against.");
+        }
+        if (!game.getStoredValue("agendaRepGov").isEmpty()) {
+            for (Player p2 : game.getRealPlayers()) {
+                if (game.getStoredValue("agendaRepGov").contains(p2.getFaction())) {
+                    for (String planet : p2.getPlanets()) {
+                        Planet p = game.getPlanetsInfo().get(planet);
+                        if (p != null && p.getPlanetTypes().contains("cultural")) {
+                            p2.exhaustPlanet(planet);
+                        }
+                    }
+                }
+            }
+            game.setStoredValue("agendaRepGov", "");
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "# Exhausted all cultural planets of those who voted against on that one agenda");
+        }
+        if (game.isFowMode()) {
+            MessageHelper.sendMessageToChannel(event.getMessageChannel(), "Pinged speaker to pick a strategy card.");
+        }
+        Player speaker;
+        if (game.getPlayer(game.getSpeaker()) != null) {
+            speaker = game.getPlayers().get(game.getSpeaker());
+        } else {
+            MessageHelper.sendMessageToChannel(event.getMessageChannel(), "Speaker not found. Can't proceed");
+            return;
+        }
+        String message = speaker.getRepresentation(true, true) + " UP TO PICK SC\n";
+        game.updateActivePlayer(speaker);
+        game.setPhaseOfGame("strategy");
+        String pickSCMsg = "Use buttons to pick a strategy card.";
+        if (game.getLaws().containsKey("checks") || game.getLaws().containsKey("absol_checks")) {
+            pickSCMsg = "Use buttons to pick the strategy card you want to give someone else.";
+        }
+        ButtonHelperAbilities.giveKeleresCommsNTg(game, event);
+        game.setStoredValue("startTimeOfRound" + game.getRound() + "Strategy", new Date().getTime() + "");
+        if (game.isFowMode()) {
+            if (!game.isHomebrewSCMode()) {
+                MessageHelper.sendMessageToChannelWithButtons(speaker.getPrivateChannel(), message + pickSCMsg, Helper.getRemainingSCButtons(event, game, speaker));
+            } else {
+                MessageHelper.sendPrivateMessageToPlayer(speaker, game, message);
+            }
+        } else {
+            MessageHelper.sendMessageToChannelWithButtons(game.getMainGameChannel(), message + pickSCMsg,
+                Helper.getRemainingSCButtons(event, game, speaker));
+        }
+        if (!game.isFowMode()) {
+            ButtonHelper.updateMap(game, event, "Start of Strategy Phase For Round #" + game.getRound());
+        }
+        for (Player player2 : game.getRealPlayers()) {
+            if (player2.getActionCards() != null && player2.getActionCards().containsKey("summit")) {
+                MessageHelper.sendMessageToChannel(player2.getCardsInfoThread(), player2.getRepresentation(true, true) + "Reminder this is the window to play Summit.");
+            }
+            for (String pn : player2.getPromissoryNotes().keySet()) {
+                if (!player2.ownsPromissoryNote("scepter") && "scepter".equalsIgnoreCase(pn)) {
+                    PromissoryNoteModel promissoryNote = Mapper.getPromissoryNote(pn);
+                    Player owner = game.getPNOwner(pn);
+                    Button transact = Buttons.green("resolvePNPlay_" + pn, "Play " + promissoryNote.getName(), owner.getFactionEmoji());
+                    List<Button> buttons = new ArrayList<>();
+                    buttons.add(transact);
+                    buttons.add(Buttons.red("deleteButtons", "Decline"));
+                    String cyberMessage = player2.getRepresentation(true, true) + " reminder this is the window to play Mahact PN if you want (button should work)";
+                    MessageHelper.sendMessageToChannelWithButtons(player2.getCardsInfoThread(), cyberMessage, buttons);
+                    if (!game.isFowMode()) {
+                        MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "You should all pause for a potential mahact PN play here if you think it relevant");
+                    }
+                }
+            }
+        }
+        if (game.getTile("SIG02") != null && !game.isFowMode()) {
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "Please destroy all units in the pulsar.");
+        }
+        if ("action_deck_2".equals(game.getAcDeckID())) {
+            handleStartOfStrategyForAcd2(game);
+        }
+    }
+
+    private static void handleStartOfStrategyForAcd2(Game game) {
+        boolean deflectionDiscarded = game.isACInDiscard("Deflection");
+        boolean revolutionDiscarded = game.isACInDiscard("Revolution");
+        StringJoiner stringJoiner = new StringJoiner(" and ");
+        if (!deflectionDiscarded)
+            stringJoiner.add("*Deflection*");
+        if (!revolutionDiscarded)
+            stringJoiner.add("*Revolution*");
+        String acd2Shenanigans;
+        if (stringJoiner.length() > 0) {
+            acd2Shenanigans = "This is the window for " + stringJoiner + "! " + game.getPing();
+            handleStartOfStrategyForAcd2Player(game);
+        } else {
+            acd2Shenanigans = "*Deflection* and *Revolution* are in the discard pile. Feel free to move forward.";
+        }
+        MessageHelper.sendMessageToChannel(game.getMainGameChannel(), acd2Shenanigans);
+    }
+
+    private static void handleStartOfStrategyForAcd2Player(Game game) {
+        for (Player player : game.getRealPlayers()) {
+            if (player.getActionCards().containsKey("deflection")) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation(true, true) + "Reminder this is the window to play Deflection.");
+            }
+            if (player.getActionCards().containsKey("revolution")) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation(true, true) + "Reminder this is the window to play Revolution.");
+            }
+        }
+    }
+
+    public static void startStatusHomework(GenericInteractionCreateEvent event, Game game) {
+        game.setPhaseOfGame("statusHomework");
+        game.setStoredValue("startTimeOfRound" + game.getRound() + "StatusHomework", new Date().getTime() + "");
+        // first do cleanup if necessary
+        int playersWithSCs = 0;
+        for (Player player : game.getRealPlayers()) {
+            if (player.getSCs() != null && player.getSCs().size() > 0 && !player.getSCs().contains(0)) {
+                playersWithSCs++;
+            }
+        }
+
+        if (playersWithSCs > 0) {
+            new Cleanup().runStatusCleanup(game);
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), game.getPing() + " **Status Cleanup Run!**");
+            if (!game.isFowMode()) {
+                DisplayType displayType = DisplayType.map;
+                MapGenerator.saveImage(game, displayType, event).thenAccept(fileUpload -> MessageHelper.sendFileUploadToChannel(game.getActionsChannel(), fileUpload));
+            }
+        }
+
+        for (Player player : game.getRealPlayers()) {
+            Leader playerLeader = player.getLeader("naaluhero").orElse(null);
+
+            if (game.getRound() < 4) {
+                String preferences = "";
+                for (Player p2 : game.getRealPlayers()) {
+                    if (p2 == player) {
+                        continue;
+                    }
+                    String old = game.getStoredValue(p2.getUserID() + "anonDeclare");
+                    if (!old.isEmpty() && !old.toLowerCase().contains("strong")) {
+                        preferences = preferences + old + "; ";
+                    }
+                }
+                if (!preferences.isEmpty()) {
+                    preferences = preferences.substring(0, preferences.length() - 2);
+                    preferences = player.getRepresentation() + " this is a reminder that at the start of the game, your fellow players stated a preference for the following environments:\n" +
+                        preferences + "\nYou are under no special obligation to abide by that preference, but it may be a nice thing to keep in mind as you play";
+                    MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), preferences);
+                }
+            }
+
+            if (player.hasLeader("naaluhero") && player.getLeaderByID("naaluhero").isPresent() && playerLeader != null && !playerLeader.isLocked()) {
+                List<Button> buttons = new ArrayList<>();
+                buttons.add(Buttons.green("naaluHeroInitiation", "Play Naalu Hero", Emojis.NaaluHero));
+                buttons.add(Buttons.red("deleteButtons", "Decline"));
+                MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), player.getRepresentation() + " Reminder this is the window to play The Oracle, the Naalu Hero. You may use the buttons to start the process.", buttons);
+            }
+            if (player.getRelics() != null && player.hasRelic("mawofworlds") && game.isCustodiansScored()) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation() + " Reminder this is the window to do Maw of Worlds, after you do your status homework things. Maw of worlds is technically start of agenda, but can be done now for efficiency");
+                MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), player.getRepresentation() + " You may use these buttons to resolve Maw Of Worlds.", ButtonHelper.getMawButtons());
+            }
+            if (player.getRelics() != null && player.hasRelic("twilight_mirror") && game.isCustodiansScored()) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation() + " Reminder this is the window to do Twilight Mirror");
+                List<Button> playerButtons = new ArrayList<>();
+                playerButtons.add(Buttons.green("resolveTwilightMirror", "Purge Twilight Mirror", Emojis.Relic));
+                playerButtons.add(Buttons.red("deleteButtons", "Decline"));
+                MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), player.getRepresentation() + " You may use these buttons to resolve Twilight Mirror.", playerButtons);
+            }
+            if (player.getRelics() != null && player.hasRelic("emphidia")) {
+                for (String pl : player.getPlanets()) {
+                    Tile tile = game.getTile(AliasHandler.resolveTile(pl));
+                    if (tile == null) {
+                        continue;
+                    }
+                    UnitHolder unitHolder = tile.getUnitHolders().get(pl);
+                    if (unitHolder != null && unitHolder.getTokenList() != null && unitHolder.getTokenList().contains("attachment_tombofemphidia.png")) {
+                        MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation() + "Reminder this is the window to purge Crown of Emphidia if you want to.");
+                        MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(), player.getRepresentation() + " You may use these buttons to resolve Crown of Emphidia.", ButtonHelper.getCrownButtons());
+                    }
+                }
+            }
+
+            if (player.getActionCards() != null && player.getActionCards().containsKey("stability")) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation() + "Reminder this is the window to play Political Stability.");
+            }
+
+            if (player.getActionCards() != null && player.getActionCards().containsKey("abs") && game.isCustodiansScored()) {
+                MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), player.getRepresentation() + "Reminder this is the window to play Ancient Burial Sites.");
+            }
+
+            for (String pn : player.getPromissoryNotes().keySet()) {
+                if (!player.ownsPromissoryNote("ce") && "ce".equalsIgnoreCase(pn)) {
+                    String cyberMessage = "# " + player.getRepresentation(true, true) + " reminder to use cybernetic enhancements!";
+                    MessageHelper.sendMessageToChannel(player.getCardsInfoThread(), cyberMessage);
+                }
+            }
+        }
+        String message2 = "Resolve status homework using the buttons. \n ";
+        game.setCurrentACDrawStatusInfo("");
+        Button draw1AC = Buttons.green("drawStatusACs", "Draw Status Phase ACs", Emojis.ActionCard);
+        Button getCCs = Buttons.green("redistributeCCButtons", "Redistribute, Gain, & Confirm CCs").withEmoji(Emoji.fromFormatted("🔺"));
+        Button yssarilPolicy = null;
+        for (Player player : game.getRealPlayers()) {
+            if (ButtonHelper.isPlayerElected(game, player, "minister_policy") && player.hasAbility("scheming")) {
+                yssarilPolicy = Buttons.gray(player.getFinsFactionCheckerPrefix() + "_yssarilMinisterOfPolicy", "Draw Minister of Policy AC", Emojis.Yssaril);
+            }
+        }
+        boolean custodiansTaken = game.isCustodiansScored();
+        Button passOnAbilities;
+        if (custodiansTaken) {
+            passOnAbilities = Buttons.red("pass_on_abilities", "Ready For Agenda");
+            message2 = message2 + "This is the moment when you should resolve: \n- Political Stability \n- Ancient Burial Sites\n- Maw of Worlds \n- The Oracle, the Naalu hero\n- Crown of Emphidia";
+        } else {
+            passOnAbilities = Buttons.red("pass_on_abilities", "Ready For Strategy Phase");
+            message2 = message2 + "Ready For Strategy Phase means you are done playing/passing on: \n- Political Stability \n- Summit \n- Manipulate Investments ";
+        }
+        List<Button> buttons = new ArrayList<>();
+        buttons.add(draw1AC);
+        buttons.add(getCCs);
+        buttons.add(passOnAbilities);
+        if (yssarilPolicy != null) {
+            buttons.add(yssarilPolicy);
+        }
+        MessageHelper.sendMessageToChannelWithButtons(game.getMainGameChannel(), message2, buttons);
+        if (game.isFowMode()) {
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "Remember to click Ready for " + (custodiansTaken ? "Agenda" : "Strategy Phase") + " when done with homework!");
+        }
+    }
+
+    public static void startActionPhase(GenericInteractionCreateEvent event, Game game) {
+        boolean isFowPrivateGame = FoWHelper.isPrivateGame(game, event);
+        String msg;
+        game.setStoredValue("willRevolution", "");
+        game.setPhaseOfGame("action");
+        String msgExtra = "";
+        boolean allPicked = true;
+        Player privatePlayer = null;
+        Collection<Player> activePlayers = game.getPlayers().values().stream()
+            .filter(Player::isRealPlayer)
+            .toList();
+        Player nextPlayer = null;
+        int lowestSC = 100;
+        for (Player p2 : game.getRealPlayers()) {
+            ButtonHelperActionCards.checkForAssigningCoup(game, p2);
+            if (game.getStoredValue("Play Naalu PN") != null
+                && game.getStoredValue("Play Naalu PN").contains(p2.getFaction())) {
+                if (!p2.getPromissoryNotesInPlayArea().contains("gift")
+                    && p2.getPromissoryNotes().containsKey("gift")) {
+                    PlayPN.resolvePNPlay("gift", p2, game, event);
+                }
+            }
+        }
+        for (Player player_ : activePlayers) {
+            int playersLowestSC = player_.getLowestSC();
+            String scNumberIfNaaluInPlay = game.getSCNumberIfNaaluInPlay(player_,
+                Integer.toString(playersLowestSC));
+            if (scNumberIfNaaluInPlay.startsWith("0/")) {
+                nextPlayer = player_; // no further processing, this player has the 0 token
+                break;
+            }
+            if (playersLowestSC < lowestSC) {
+                lowestSC = playersLowestSC;
+                nextPlayer = player_;
+            }
+        }
+
+        // INFORM FIRST PLAYER IS UP FOR ACTION
+        if (nextPlayer != null) {
+            msgExtra += " " + nextPlayer.getRepresentation() + " is up for an action";
+            privatePlayer = nextPlayer;
+            game.updateActivePlayer(nextPlayer);
+            if (game.isFowMode()) {
+                FoWHelper.pingAllPlayersWithFullStats(game, event, nextPlayer, "started turn");
+            }
+            ButtonHelperFactionSpecific.resolveMilitarySupportCheck(nextPlayer, game);
+
+            game.setPhaseOfGame("action");
+        }
+
+        msg = "";
+        MessageHelper.sendMessageToChannel(nextPlayer.getCorrectChannel(), msg);
+        if (isFowPrivateGame) {
+            msgExtra = "Start phase command run";
+            String fail = "User for next faction not found. Report to ADMIN";
+            String success = "The next player has been notified";
+            MessageHelper.sendPrivateMessageToPlayer(privatePlayer, game, event, msgExtra, fail, success);
+            if (privatePlayer == null)
+                return;
+            msgExtra = "" + privatePlayer.getRepresentation(true, true) + " UP NEXT";
+            game.updateActivePlayer(privatePlayer);
+
+            if (!allPicked) {
+                game.setPhaseOfGame("strategy");
+                MessageHelper.sendMessageToChannelWithButtons(privatePlayer.getPrivateChannel(), "Use buttons to pick a strategy card.", Helper.getRemainingSCButtons(event, game, privatePlayer));
+            } else {
+
+                MessageHelper.sendMessageToChannelWithButtons(privatePlayer.getPrivateChannel(), msgExtra + "\n Use Buttons to do turn.", TurnStart.getStartOfTurnButtons(privatePlayer, game, false, event));
+
+                if (privatePlayer.getStasisInfantry() > 0) {
+                    if (ButtonHelper.getPlaceStatusInfButtons(game, privatePlayer).size() > 0) {
+                        MessageHelper.sendMessageToChannelWithButtons(privatePlayer.getCorrectChannel(), "Use buttons to revive infantry. You have " + privatePlayer.getStasisInfantry() + " infantry left to revive.", ButtonHelper.getPlaceStatusInfButtons(game, privatePlayer));
+                    } else {
+                        privatePlayer.setStasisInfantry(0);
+                        MessageHelper.sendMessageToChannel(privatePlayer.getCorrectChannel(), privatePlayer.getRepresentation() + " You had infantry II to be revived, but the bot couldn't find planets you own in your HS to place them, so per the rules they now disappear into the ether");
+                    }
+                }
+            }
+
+        } else {
+            MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "All players have picked a strategy card.");
+            if (game.isShowBanners()) {
+                MapGenerator.drawPhaseBanner("action", game.getRound(), game.getActionsChannel());
+            }
+            ListTurnOrder.turnOrder(event, game);
+            if (!msgExtra.isEmpty()) {
+                MessageHelper.sendMessageToChannel(game.getMainGameChannel(), msgExtra);
+                if (privatePlayer == null) {
+                    MessageHelper.sendMessageToChannel(game.getMainGameChannel(), "Could not find player.");
+                    return;
+                }
+                MessageHelper.sendMessageToChannelWithButtons(game.getMainGameChannel(), "\n Use Buttons to do turn.", TurnStart.getStartOfTurnButtons(privatePlayer, game, false, event));
+
+                if (privatePlayer.getStasisInfantry() > 0) {
+                    if (ButtonHelper.getPlaceStatusInfButtons(game, privatePlayer).size() > 0) {
+                        MessageHelper.sendMessageToChannelWithButtons(privatePlayer.getCorrectChannel(), "Use buttons to revive infantry. You have " + privatePlayer.getStasisInfantry() + " infantry left to revive.", ButtonHelper.getPlaceStatusInfButtons(game, privatePlayer));
+                    } else {
+                        privatePlayer.setStasisInfantry(0);
+                        MessageHelper.sendMessageToChannel(privatePlayer.getCorrectChannel(), privatePlayer.getRepresentation() + " You had infantry II to be revived, but the bot couldn't find planets you own in your HS to place them, so per the rules they now disappear into the ether.");
+                    }
+                }
+            }
+        }
+        for (Player p2 : game.getRealPlayers()) {
+            List<Button> buttons = new ArrayList<>();
+            if (p2.hasTechReady("qdn") && p2.getTg() > 2 && p2.getStrategicCC() > 0) {
+                buttons.add(Buttons.green("startQDN", "Use Quantum Datahub Node", Emojis.CyberneticTech));
+                buttons.add(Buttons.red("deleteButtons", "Decline"));
+                MessageHelper.sendMessageToChannelWithButtons(p2.getCorrectChannel(), p2.getRepresentation(true, true) + " you have the opportunity to use QDN", buttons);
+            }
+            buttons = new ArrayList<>();
+            if (ButtonHelper.isPlayerElected(game, p2, "arbiter")) {
+                buttons.add(Buttons.green("startArbiter", "Use Imperial Arbiter", Emojis.Agenda));
+                buttons.add(Buttons.red("deleteButtons", "Decline"));
+                MessageHelper.sendMessageToChannelWithButtons(p2.getCorrectChannel(), p2.getRepresentation(true, true) + " you have the opportunity to use Imperial Arbiter", buttons);
+            }
         }
     }
 }
